requirements.txt
requests
beautifulsoup4
aiohttp
crawl4ai
python-dotenv
